refactor: split prompts module, extract oversized helpers, fix silent exception swallows
PR 5 — Exception handling + structural extractions (behavior-preserving):

- Fix ~4 true silent-swallow exception handlers: add log.debug() in
  chain.py (APOC fallback, chain-path error), _state.py (agent-state
  hash), builder.py (compat-shim catch). No control-flow changes.

- Split decepticon/agents/prompts/__init__.py (477→78 lines) into:
  - builder.py: prompt assembly engine (build_system_prompt, etc.)
  - registry.py: language policy, LANG_COUNTRY_MAP, TRANSLATION_MAP
  __init__.py becomes a re-export shim; all public names preserved.

- Extract private helpers from 4 oversized functions:
  - llm/factory.py: 13 helpers for credential detection & resolver
  - sandbox_kernel/tmux.py: helpers for pane init and config setup
  - tools/research/scanner_tools.py: _build_candidates helper
  - tools/skills.py: _read_frontmatter, _write_loaded_skill helpers

All existing tests remain green. No public API or docstring changes.
